requirements.txt
streamlit
google-genai
pypdf2